CI/CD, airflow, refactor code
