+++ airflow/logs/scheduler/latest
@@ -1,0 +1,1 @@
+2024-07-27
