--- transform/sql_script/change_data_type.sql
@@ -1,54 +1,0 @@
-ALTER TABLE [dbo].bronze_layer
-MODIFY Type TEXT,
-MODIFY DaysForShipping INT,
-MODIFY DaysForShipment INT,
-MODIFY BenefitPerOrder DOUBLE,
-MODIFY SalesPerCustomer DOUBLE,
-MODIFY DeliveryStatus TEXT,
-MODIFY LateDeliveryRisk INT,
-MODIFY CategoryId INT,
-MODIFY CategoryName TEXT,
-MODIFY CustomeCity TEXT,
-MODIFY CustomerCountry TEXT,
-MODIFY CustomerEmail TEXT,
-MODIFY CustomerFname TEXT,
-MODIFY CustomerId INT,
-MODIFY CustomerLname TEXT,
-MODIFY CustomerPassword TEXT,
-MODIFY CustomerSegment TEXT,
-MODIFY CustomerState TEXT,
-MODIFY CustomerStreet TEXT,
-MODIFY CustomerZipcode INT,
-MODIFY DepartmentId INT,
-MODIFY DepartmentName TEXT,
-MODIFY Latitude DOUBLE,
-MODIFY Longitude DOUBLE,
-MODIFY Market TEXT,
-MODIFY OrderCity TEXT,
-MODIFY OrderCountry TEXT,
-MODIFY OrderCustomerId INT,
-MODIFY Orderdate TEXT,
-MODIFY OrderId INT,
-MODIFY OrderItemCardprodId INT,
-MODIFY OrderItemDiscount DOUBLE,
-MODIFY OrderItemDiscountRate DOUBLE,
-MODIFY OrderItemId INT,
-MODIFY OrderItemProductPrice DOUBLE,
-MODIFY OrderItemProfitRatio DOUBLE,
-MODIFY OrderItemQuantity INT,
-MODIFY Sales DOUBLE,
-MODIFY OrderItemTotal DOUBLE,
-MODIFY OrderProfitPerOrder DOUBLE,
-MODIFY OrderRegion TEXT,
-MODIFY OrderState TEXT,
-MODIFY OrderStatus TEXT,
-MODIFY OrderZipcode TEXT,
-MODIFY ProductCardId INT,
-MODIFY ProductCategoryId TEXT,
-MODIFY ProductDescription TEXT,
-MODIFY ProductImage TEXT,
-MODIFY ProductName TEXT,
-MODIFY ProductPrice TEXT,
-MODIFY ProductStatus TEXT,
-MODIFY Shippingdate TEXT,
-MODIFY ShippingMode TEXT;
